IMPALA-1619: Support 64-bit allocations.
This change extends MemPool, FreePool and StringBuffer to support
64-bit allocations, fixes a bug in decompressor and extends various
places in the code to support 64-bit allocation sizes. With this
change, the text scanner can now decompress compressed files larger
than 1GB.

Note that the UDF interfaces FunctionContext::Allocate() and
FunctionContext::Reallocate() still use 32-bit for the input
argument to avoid breaking compatibility.

In addition, the byte size of a tuple is still assumed to be
within 32-bit. If it needs to be upgraded to 64-bit, it will be
done in a separate change.

A new test has been added to test the decompression of a 2GB
snappy block compressed text file.

Reviewed-on: http://gerrit.cloudera.org:8080/3575
Tested-by: Internal Jenkins
(cherry picked from commit 80a92d8afd002667df34aa194e1d5333b8b8ae24)
(cherry picked from commit 0dada13938f41b6e40c101861c96c495b29c5f29)
Reviewed-on: http://gerrit.sjc.cloudera.com:8080/16974
Tested-by: Anuj Phadke <aphadke@cloudera.com>

+++ be/src/runtime/collection-value-builder.h
@@ -29,7 +29,7 @@ class CollectionValueBuilder {
   static const int DEFAULT_INITIAL_TUPLE_CAPACITY = 4;
 
   CollectionValueBuilder(CollectionValue* coll_value, const TupleDescriptor& tuple_desc,
-      MemPool* pool, int initial_tuple_capacity = DEFAULT_INITIAL_TUPLE_CAPACITY)
+      MemPool* pool, int64_t initial_tuple_capacity = DEFAULT_INITIAL_TUPLE_CAPACITY)
     : coll_value_(coll_value),
       tuple_desc_(tuple_desc),
       pool_(pool) {
